==> 1.econdec-12_merge.ipynb <==
# ED-1 & ED-2 Merge
`1.econdec-12_merge`

Combine task-specific datasets from EconDec-1 & EconDec-2 and situate for homogenization & merging with dataset from EconDec-3

import os
import pandas as pd
from datetime import datetime

date = datetime.today().strftime('%y%m%d')

verbose = False

source_dir = os.path.abspath(os.path.join('..','sourcedata'))
output_dir = os.path.abspath(os.path.join('..','derivatives','00.allsub'))
if not os.path.isdir(output_dir): os.mkdir(output_dir)

frames={'main':[],'frac':[],'face':[]}
subs_missing_data=[]
for dataset in ['ds1','ds2']:
    dataset_dir = os.path.join(source_dir, dataset)
    for sub in os.listdir(dataset_dir):
        if verbose: print(sub)
        sub_dir = os.path.join(dataset_dir, sub)
        for task in ['main','frac','face']:
            infile_name = sub +'_task-'+ task +'_beh.xlsx'
            infile_path = os.path.join(sub_dir, infile_name)
            try:
                frames[task].append(pd.read_excel(infile_path))
                if verbose: print('[Extract]', fname)
            except IOError as e:
                if verbose: print(e)
                subs_missing_data.append(sub)
                continue

log = '[Missing data]: ' +','.join(subs_missing_data)
print(log)

for task in ['main','frac','face']:
    outfile_name = 'econdec-12_task-' + task + '_beh_' + date + '.csv'
    outfile_path = os.path.join(output_dir, outfile_name)
    pd.concat(frames[task]).to_csv(outfile_path, index=False)
    print('[Output]',outfile_path)
