==> 1.econdec-3_convert.ipynb <==
# EconDec-3 Conversion
`1.econdec-3_convert`

Concatenate converted Eye-study files and situate them for downstream homogenization with Study-1 and Study-2

import os
import pandas as pd
from datetime import datetime

verbose = False

source_dir = os.path.join('..','sourcedata','ds3')
output_dir = os.path.join('..','derivatives','00.allsub')

date = datetime.today().strftime("%y%m%d")

Frames = []
subs = []
for s in os.listdir(source_dir):
    if s.startswith('sub-'):
        subs.append(s)
        sub_dir = os.path.join(source_dir,s)
        file_path = os.path.join(sub_dir,'RESULTS_FILE.txt')
        if verbose: print('Reading: '+file_path)
        file = pd.read_csv(file_path, delimiter='\t')
        Frames.append(file)

print('[Collected]:',','.join(subs))

output_frame = pd.concat(Frames)
output_path = os.path.join(output_dir,'econdec-3_task-all_beh_' + date + '.csv')
output_frame.to_csv(output_path, index=False)
