==> 1.study-one_merge.ipynb <==
import os
import pandas as pd

source_dir = os.path.abspath(os.path.join('..','sourcedata'))
output_dir = os.path.abspath(os.path.join('..','derivatives','00.allsub'))
try: os.mkdir(output_dir)
except WindowsError as e: print(e)

datasets = ['ds1','ds2']
frames=[]
for ds in datasets:
    dspath = os.path.join(source_dir,ds)
    for sub in os.listdir(dspath):
        fpath = os.path.join(dspath,sub,sub+'_task-main_beh.xlsx')
        try:
            frames.append(pd.read_excel(fpath))
        except IOError as e:
            print(e)
            continue

output_path = os.path.join(output_dir,'study-one_task-main_beh.csv')
pd.concat(frames).to_csv(output_path,index=False)

datasets = ['ds1','ds2']
frames=[]
for ds in datasets:
    dspath = os.path.join(source_dir,ds)
    for sub in os.listdir(dspath):
        fpath = os.path.join(dspath,sub,sub+'_task-frac_beh.xlsx')
        try:
            frames.append(pd.read_excel(fpath))
        except IOError as e:
            print(e)
            continue

output_path = os.path.join(output_dir,'study-one_task-frac_beh.csv')
pd.concat(frames).to_csv(output_path,index=False)

datasets = ['ds1','ds2']
frames=[]
for ds in datasets:
    dspath = os.path.join(source_dir,ds)
    for sub in os.listdir(dspath):
        fpath = os.path.join(dspath,sub,sub+'_task-face_beh.xlsx')
        try:
            frames.append(pd.read_excel(fpath))
        except IOError as e:
            print(e)
            continue

output_path = os.path.join(output_dir,'study-one_task-face_beh.csv')
pd.concat(frames).to_csv(output_path,index=False)
