==> 2.homogenize.ipynb <==
Realign and merge converted data from eye-tracking into the concatenated data from studies 1 and 2. Differentiate, import, and reassociate memory data into the main-task trialwise dataset.

import os
import pandas as pd

derivs_dir = os.path.join('..','derivatives')
allsub_dir = os.path.join(derivs_dir,'00.allsub')

def dfsplit(input_frame, column_criterion, value_criterion, operation_flag = '=='):
    import operator
    ops = {'>' : operator.gt, '<' : operator.lt, '==': operator.eq,
           '!=': operator.ne, '<=': operator.le, '>=': operator.ge}
    op = ops[operation_flag]
    output_frame = input_frame[op(input_frame[column_criterion],value_criterion)]
    return output_frame

def smooth_columns(input_frame):
    column_labels = list(input_frame.columns)
    input_frame.columns = [x.lower().replace('_','') for x in column_labels]
    return input_frame

eye_fpath = os.path.join(allsub_dir,'study-eye_task-all_beh.csv')
eye_frame = pd.read_csv(eye_fpath)
main_eye_frame = smooth_columns(dfsplit(eye_frame, 'Phase', 'Main Task'))
frac_eye_frame = smooth_columns(dfsplit(eye_frame, 'Phase', 'Fract'))
face_eye_frame = smooth_columns(dfsplit(eye_frame, 'Phase', 'Face'))

main_fpath = os.path.join(allsub_dir,'study-one_task-main_beh.csv')
frac_fpath = os.path.join(allsub_dir,'study-one_task-frac_beh.csv')
face_fpath = os.path.join(allsub_dir,'study-one_task-face_beh.csv')

main_frame = smooth_columns(pd.read_csv(main_fpath))
frac_frame = smooth_columns(pd.read_csv(frac_fpath))
face_frame = smooth_columns(pd.read_csv(face_fpath))

# Exclude bad subjects

exclusions=[
    117,165,183,189,
    201,217,247,254,265,270,278,283,284,
    300,307,323
]

#### Study-one

main_frame = main_frame[~main_frame['subjnum'].isin(exclusions)]

len(main_frame.subjnum.unique())

#### Study-eye

main_eye_frame = main_eye_frame[~main_eye_frame['originalparticipant'].isin(exclusions)]

len(main_eye_frame.originalparticipant.unique())

#### Output

exclusions_dir = os.path.join(derivs_dir,'01.exclusions')
try:
    os.mkdir(exclusions_dir)
except: pass
main_frame.to_csv(os.path.join(exclusions_dir,'study-one_task-main_beh.csv'))
main_eye_frame.to_csv(os.path.join(exclusions_dir,'study-eye_task-all_beh.csv'))

# Main task

## Prime study

main_frame

## Eye study

def eye_cleanup(input_frame):
    f=input_frame
    f=f[f['facekeypressed']!='.']
    f=f[f['practice']>2]
    return f

main_eye_frame = eye_cleanup(main_eye_frame)

## Main task columns

### Prime study

main_frame.columns

new_main_columns = {'face':'facepic','runnum':'block','trialnum':'trial',
               'fracrt':'choicert','fracst':'choicest',
               'facert':'outcomert','facest':'outcomest',
               'probrt':'esttaskrt','probst':'esttaskst',
               'optionchosen':'stockchosen',
               'probgood':'estimation','trueprobgood':'trueprob'}

main_frame = main_frame.rename(columns=new_main_columns)

new_main_frame = main_frame.drop(['agegroup','experimentername','date','time','trialnumbydomdist',
                                  'choicest','outcomest','esttaskst',
                                  'confidencest','stocknumber','bondnumber','genderjudgment',
                                  'totalpayout','fractalchosen','estwithinrange?','confidencert'], axis=1)

def label_study(row):
    study_label = str(row['subjnum'])[0]
    return study_label

new_main_frame['study'] = new_main_frame.apply(label_study, axis=1)
new_main_frame.columns

### Eye study

main_eye_frame.columns

new_eye_columns = {
    'originalparticipant':'subjnum','experimenter':'experimentername',
    'stockimagename':'stockpic','bondimagename':'bondpic',
    'originalblock':'block','originaltrialorder':'trial',
    #'stockfractallocationtype':'cueonleft','bondfractallocationtype':'cueonright',
    'faceimage':'facepic',
    #'selection':'stockchosen',
    'rt':'choicert',
    'estimationvalue':'estimation','trueprobability':'trueprob','accuracy':'genderjudgment',
    'confidencevalue':'confidence','facert':'outcomert'
    }

main_eye_frame = main_eye_frame.rename(columns=new_eye_columns)

new_main_eye_frame = main_eye_frame.drop([
    'practice','bankaccount','bubblefile','bondvalue','stocktext','bondtext',
    'stocktextlocation','bondtextlocation','emotionresponse','bypassed','agegroup','experimentername',
    'date','correctfractallocation','incorrectfractallocation','paymentaccuracy','phase',
    'stockfractallocation','bondfractallocation','showinstruction','gender',
    'correctfractal','incorectfractal','oldfaceequalstrue','facefile','facekeypressed',
    'originalsubjectnumber','originalparticipantnumber','originaltrialnumber','originaltrailnumber',
    'fracdomain','facedomain','fracmagnitude','facestockvalue',
    'genderjudgment'], axis=1)

new_main_eye_frame['study'] = new_main_eye_frame.apply(label_study, axis=1)

### Unified columns

#### final cleaning to put values in the same units, etc...

def clean_choicert(row):
    if int(row['study']) >= 3:
        return float(row['choicert'] * .001)
    else:
        return row['choicert']

def clean_outcomert(row):
    if int(row['study']) >= 3:
        return float(row['outcomert'] * .001)
    else:
        return row['outcomert']

def clean_esttaskrt(row):
    if int(row['study']) >= 3:
        return float(row['esttaskrt'] * .001)
    else:
        return row['esttaskrt']

def clean_stockchosen(row):
    if int(row['study']) >= 3:
        if row['selection'] == 1:
            # chose option left
            if row['stockfractallocationtype'] == 'L':
                # stock was on left 
                # row['stockchosen'] = 1
                return 1
            elif row['stockfractallocationtype'] == 'R':
                # stock was on right
                # row['stockchosen'] = 0
                return 0
        elif row['selection'] == 0:
            # chose option right
            if row['stockfractallocationtype'] == 'R':
                # stock was on right
                # row['stockchosen'] = 1
                return 1
            elif row['stockfractallocationtype'] == 'L':
                # stock was on left
                # row['stockchosen'] = 0
                return 0
    else:
        if row['stockchosen'] == 'stock':
            return 1
        elif row['stockchosen'] == 'bond':
            return 0

def clean_bondpic(row):
    return os.path.split(row['bondpic'])[1]

def clean_stockpic(row):
    return os.path.split(row['stockpic'])[1]

unified_main_frame = pd.concat([new_main_frame,new_main_eye_frame])

unified_main_frame['choicert'] = unified_main_frame.apply(clean_choicert, axis=1)
unified_main_frame['outcomert'] = unified_main_frame.apply(clean_outcomert, axis=1)
unified_main_frame['stockchosen'] = unified_main_frame.apply(clean_stockchosen, axis=1)
unified_main_frame['bondpic'] = unified_main_frame.apply(clean_bondpic, axis=1)
unified_main_frame['stockpic'] = unified_main_frame.apply(clean_stockpic, axis=1)
len(unified_main_frame)

# Fractal Memory

## Prime study

def clean_paths(row):
    return os.path.basename(row['oldfractal'])

frac_frame['oldfractal'] = frac_frame.apply(clean_paths, axis=1)

frac_lil_frame = frac_frame[['subjectid','oldfractal','judgment']].sort_values(['subjectid','oldfractal'])

frac_lil_bond_frame = frac_lil_frame.rename(columns={
    'subjectid':'subjnum','oldfractal':'bondpic','judgment':'bondmem'
})

frac_lil_stock_frame = frac_lil_frame.rename(columns={
    'subjectid':'subjnum','oldfractal':'stockpic','judgment':'stockmem'
})

# Eye study

def clean_selection(row):
    if row['correctfractallocation'] == '(1355, 540)':
        # correct fractal was on right
        return int(row['selection'])-4
    elif row['correctfractallocation'] == '(565, 540)':
        # correct fractal was on left
        return -int(row['selection'])+5

frac_eye_lil_frame = frac_eye_frame[['originalparticipant','correctfractal','selection','correctfractallocation']]
frac_eye_lil_frame['selection'] = frac_eye_lil_frame.apply(clean_selection, axis=1)

frac_eye_lil_bond_frame = frac_eye_lil_frame.rename(columns={
    'originalparticipant':'subjnum',
    'correctfractal':'bondpic',
    'selection':'bondmem'
}).drop(columns='correctfractallocation')

frac_eye_lil_stock_frame = frac_eye_lil_frame.rename(columns={
    'originalparticipant':'subjnum',
    'correctfractal':'stockpic',
    'selection':'stockmem'
}).drop(columns='correctfractallocation')

frac_lil_bond_frame = pd.concat([frac_lil_bond_frame,frac_eye_lil_bond_frame])
frac_lil_stock_frame = pd.concat([frac_lil_stock_frame,frac_eye_lil_stock_frame])

# Face Memory

## Prime study

face_lil_frame=face_frame[['subjectid','face','subjresp']]
face_lil_frame=face_lil_frame.rename(columns={
    'subjectid':'subjnum','face':'facepic','subjresp':'facemem'
})

face_lil_frame

## Eye study

face_eye_lil_frame=face_eye_frame[
    ['originalparticipant','facefile','selection']
].rename(columns={
    'originalparticipant':'subjnum',
    'facefile':'facepic',
    'selection':'facemem'
})

face_lil_frame=pd.concat([face_lil_frame,face_eye_lil_frame])

face_lil_frame

#### output

unified_main_frame = unified_main_frame.merge(frac_lil_bond_frame,how='left')
unified_main_frame = unified_main_frame.merge(frac_lil_stock_frame,how='left')
unified_main_frame = unified_main_frame.merge(face_lil_frame,how='left')
# unified_main_frame[['subjnum','stockpic','bondpic','stockmem','bondmem']]

trials=[]
for s in range(len(unified_main_frame.subjnum.unique())):
    for t in range(1,73):
        trials.append(t)

blocks=[]
for s in range(len(unified_main_frame.subjnum.unique())):
    for b in range(1,13):
        for x in range(6):
            blocks.append(b)

assert len(blocks) == len(trials)
assert len(trials) == len(unified_main_frame)

unified_main_frame['trial'] = pd.Series(trials)
unified_main_frame['block'] = pd.Series(blocks)

homog_dir=os.path.join(derivs_dir,'02.homogenized')
try: os.mkdir(homog_dir)
except OSError as e:
    print(e)

fpath = os.path.join(homog_dir,'econdec-full_task-main_beh.csv')    
unified_main_frame.to_csv(fpath, index=False)

len(unified_main_frame.subjnum.unique())

final_columns=['study','subjnum','trial','block','domain','dom',
               'estimation','trueprob','estdiff','valestdiff','valestdiffvalid',
               'choicert','choicerta3sd','choicerti3sd','choicemed12v3','choicemed123'
               'esttaskrt','esttaskrta3sd','esttaskrti3sd',
               'outcomert','outcomerta3sd','outcomerti3sd','outcomemed12','outcomemed123'
               'stockchosen','waschoiceoptimal','optimalchoiceshouldhavebeen',
               'magnitude','stockvalue','absstockval','b4choiceprobability',
               'stockpic','bondpic','facepic','stockmemresp','bondmemresp',
               'studymedchoice','studysplitchoice','studymedoutcome','studysplitoutcome',
               'primemedchoice','primesplitchoice','primemedoutcome','primesplitoutcome']
