==> 4.transform_trial_bin_split_znorm.ipynb <==
# Binarization, Median-splits, Z-norming
`4.trial_bin_split_znorm`

Binarize, apply median splits, mean-centering, and other secondary trial-level variant variables.

### Setup: Modules, Directories, & Data Import

import os
import pandas as pd
import numpy as np

gaze_dir = os.path.join('..','derivatives','03.gaze-import')
bins_dir = os.path.join('..','derivatives','04.binarized')
try: os.mkdir(bins_dir)
except OSError as e:
    print(e)

file = os.path.join(gaze_dir,'econdec-full_task-all_eye.csv')
data = pd.read_csv(file)

columns=[
    'study','subjnum','block','trial','domain',#'binary-domain',
    'estimation','trueprob',
    #'val-estdiff','val-estdiff-valid','val-estdiff-valid-mc-12','val-estdiff-valid-mc-123',
    #'abs-estdiff','abs-estdiff-valid','abs-estdiff-valid-mc-12','abs-estdiff-valid-mc-123',
    'choicert',#'choicert-3sd-12','esttaskrt-3sd-123','choicert-mc-12','choicert-mc-123',
    'esttaskrt',#'esttaskrt-3sd-12','esttaskrt-3sd-123','esttaskrt-mc-12','esttaskrt-mc-123',
    'outcomert',#'outcomert-3sd-12','outcomert-3sd-123','outcomert-mc-12','esttaskrt-mc-123',
    'stockchosen',#'waschoiceoptimal','optimalchoicewas',
    'stockvalue',#'abs-stockvalue','bin-abs-stockvalue',
    #'bin-abs-stockvalue-mc-12','bin-abs-stockvalue-mc-123','b4choiceprob',
    'stockpic','bondpic','facepic','bondmem','stockmem','facemem',
    #'facemem-clean-reverse','facemem-mc-123','facemem-split-123','facemem-median-123',
    #'choice-split-12','choice-split-123','outcome-split-12','outcome-split-123',
    #'choice-median-12','choice-median-123','outcome-median-12','outcome-median-123'
    'oscillations','dwell-time','osc-rate'
]

data=data[columns]

## Median Splits

### ChoiceRT

#### choicert-median-12

data[data['study'] != 3]['choicert'].median()

#### choicert-median-123

data.groupby(['study']).median()['choicert']

### OutcomeRT

#### outcomert-median-12

data[data['study'] != 3]['outcomert'].median()

#### outcomert-median-123

data.groupby(['study']).median()['outcomert']

### FaceMem

#### clean and reverse-scale

data['facemem-clean-reverse'] = (pd.to_numeric(data['facemem'], errors='coerce') * -1) + 5

#### median-split not necessary. Mean-center?

data[data['study'] != 3]['facemem-clean-reverse'].median()

data.groupby(['study']).median()['facemem-clean-reverse']

### Output

file = os.path.join(bins_dir,'econdec-all_task-all.csv')
data.to_csv(file, index=False)
