==> 5.transform_subject_level_merge.ipynb <==
import os
import pandas as pd
import numpy as np

from _utils.clean import normalize, normalize_frame, exclude_3sd

derivs_dir = os.path.join('..','derivatives')
scales_dir = os.path.join(derivs_dir,'qualtrics','2.subscaled')
output_dir = os.path.join(derivs_dir,'05.subject-level')

try:os.mkdir(output_dir)
except WindowsError as e: print(e)

# Import data

fname=os.path.join(derivs_dir,'sub-all_task-all_VALUES.xlsx')
behav_data = pd.read_excel(fname).rename(columns={'subjnum':'ssid'})

fname=os.path.join(scales_dir,'all_subscales.csv')
scale_data = pd.read_csv(fname)

behav_data.head()

scale_data.head()

# Normalize PANAS and BISBAS subscales

keys_to_normalize = [
    'PAS','NAS','BIS','BAS_fs','BAS_rr','BAS_dr','DMS_i','DMS_r','DMS_d','DMS_s','DMS_a',
]

df = scale_data.copy(deep=True)
df.columns=[
    'ssid','zip','sleep','stress','fin_dif','PAS','NAS','BAS_dr','BAS_fs','BAS_rr','BIS',
    'DMS_i','DMS_r','DMS_d','DMS_s','DMS_a','fin_lit'
]

df['study'] = df['ssid'].astype(str).str[0]

z_keys=[]
for key in keys_to_normalize:
    z_key = 'z_'+ key
    z_keys.append(z_key)
    df[z_key] = df.apply(normalize,axis=1)

df.head()

columns = z_keys + ['ssid']
norms = df[columns]
norms.head()

### Natural Logarithm of subscales

We take the natural log of each subscale's RAW score.

*NOT* their normalized score, because we can't take the log of a negative.

ln_keys = ['ln_'+key for key in keys_to_normalize]
df[ln_keys] = df[keys_to_normalize].apply(np.log,axis=1)
df.head()

logs = df[ln_keys + ['ssid']]
logs.head()

# 3sd trial exclusions

behav_data.columns

df=behav_data[['ssid','block','trial','domain','estimation','trueprob-norm','waschoiceoptimal','val-estdiff-valid']].rename(
    columns={'val-estdiff-valid':'val_estdiff_valid'}
)

subj_3sd = df.groupby('ssid').std()['val_estdiff_valid'] * 3
subj_3sd.head()

subj_means = df.groupby('ssid').mean()['val_estdiff_valid']
subj_means.head()

df['valError_3sd'] = df.apply(exclude_3sd,axis=1)
df.head()

subj_means = df.groupby('ssid').mean()['valError_3sd']
subj_means = pd.DataFrame(subj_means).reset_index()
subj_means.head()

# gender-judgment trial exclusions

We don't have the gender judgment data in the source file for this notebook, but this is where I want to exclude trials in which the gender judgment was wrong. Have to go further upstream to include the gender judgment in this source file, then apply the gender judgment exclusion here.

# subject-level means

optimal_choice_freq = df.groupby('ssid').mean().reset_index()[['ssid','waschoiceoptimal']]
optimal_choice_freq.head()

domain_means = df.groupby(['ssid','domain']).mean().reset_index()
domain_means.head()

gain_ave_val_error = domain_means[domain_means['domain'] == 'GAIN'][['ssid','valError_3sd']]
gain_ave_val_error = gain_ave_val_error.set_index('ssid')

loss_ave_val_error = domain_means[domain_means['domain'] == 'LOSS'][['ssid','valError_3sd']]
loss_ave_val_error = loss_ave_val_error.set_index('ssid')

### Framing Normalization
We want to normalize for the way the value estimation question is framed.

We're going to multiply valError means by `1` for subjects who were estimating the probability that the stock is *good*, and multiply means by `-1` for subjects who were estimating the probability that the stock is *bad*.

100s: `*  1`

200s: `* -1`

300s: `*  1`

means_df = df.groupby('ssid').mean()[['valError_3sd']]
means_df[85:91]

nf_valerror = pd.DataFrame(means_df.apply(normalize_frame,axis=1))
nf_valerror = nf_valerror.rename(columns={0:'nf_valError'})
nf_valerror[85:91]

gain_ave_val_error[85:91]

nf_valerr_gain = pd.DataFrame(gain_ave_val_error.apply(normalize_frame,axis=1))
nf_valerr_gain = nf_valerr_gain.rename(columns={0:'nf_gainValError'})
nf_valerr_gain[85:91]

loss_ave_val_error[85:91]

nf_valerr_loss = pd.DataFrame(loss_ave_val_error.apply(normalize_frame,axis=1))
nf_valerr_loss = nf_valerr_loss.rename(columns={0:'nf_lossValError'})
nf_valerr_loss[85:91]

# output

output = pd.DataFrame({
    'ssid':list(subj_means['ssid']),
    'valError':list(subj_means['valError_3sd']),
})
output.head()

output = output.merge(optimal_choice_freq).rename(columns={'waschoiceoptimal':'optimal_choice_freq'})
output.head()

output = output.merge(gain_ave_val_error.rename(columns={'valError_3sd':'gainValError'}).reset_index())
output.head()

output = output.merge(loss_ave_val_error.rename(columns={'valError_3sd':'lossValError'}).reset_index())
output.head()

output = output.merge(nf_valerror.reset_index())
output = output.merge(nf_valerr_gain.reset_index())
output = output.merge(nf_valerr_loss.reset_index())
output[85:91]

output['valWedge'] = abs(output['gainValError'] - output['lossValError'])
output['nf_valWedge'] = abs(output['nf_gainValError'] - output['nf_lossValError'])
output.head()

output = output.merge(norms)
output.head()

output = output.merge(logs)
output.head()

fname=os.path.join(output_dir,'subject-level.csv')
#columns = ['ssid','valError','gainValError','lossValError','valWedge'] + zkeys
output.to_csv(fname,index=False)
